Move Input's components into package
