+++ Common/src/main/java/io/github/antijava/marjio/common/input/Event.java
@@ -1,4 +1,4 @@
-package io.github.antijava.marjio.common;
+package io.github.antijava.marjio.common.input;
 
 /**
  * Created by firejox on 2015/12/25.
