+++ application/src/main/java/io/github/antijava/marjio/input/AWTKeyInfo.java
@@ -1,6 +1,6 @@
 package io.github.antijava.marjio.input;
 
-import io.github.antijava.marjio.common.Key;
+import io.github.antijava.marjio.common.input.Key;
 
 import java.awt.event.KeyEvent;
 import java.util.HashMap;
