+++ application/src/main/java/io/github/antijava/marjio/input/IKeyInfo.java
@@ -1,6 +1,6 @@
 package io.github.antijava.marjio.input;
 
-import io.github.antijava.marjio.common.Key;
+import io.github.antijava.marjio.common.input.Key;
 
 /**
  * Created by firejox on 2015/12/25.
