+++ application/src/test/java/io/github/antijava/marjio/input/InputTest.java
@@ -1,7 +1,7 @@
 package io.github.antijava.marjio.input;
 
-import io.github.antijava.marjio.common.Event;
-import io.github.antijava.marjio.common.Key;
+import io.github.antijava.marjio.common.input.Event;
+import io.github.antijava.marjio.common.input.Key;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
